requirements.txt
Flask==0.12.2
Jinja2==2.10
MarkupSafe==1.0
Werkzeug==0.12.2
argparse==1.2.1
click==6.7
itsdangerous==0.24
numpy==1.13.3
scikit-surprise==1.0.4
six==1.11.0
wsgiref==0.1.2
